Add debugger module.
