+++ main.js
@@ -1,11 +1,10 @@
 import { ColorGenerator } from "./classes/ColorGenerator.js";
 import { DomManager } from "./classes/DomManager.js";
+import { Debugger } from "./modules/Debugger.js";
 
 if ("serviceWorker" in navigator) {
     navigator.serviceWorker.register("./app/serviceWorker.js");
 };
-
-const DEBUG = true;
 
 const app = (() => {
     const colorGenerator    = new ColorGenerator();
@@ -17,7 +16,7 @@ const app = (() => {
 
         domManager.init();
 
-        await colorGenerator.generatePalette(DEBUG);
+        await colorGenerator.generatePalette();
 
         // Hide loading screen
         document.getElementById("loader").classList.add("hide");
@@ -26,6 +25,9 @@ const app = (() => {
         fetch('./data/version.json')
         .then((response) => response.json())
         .then((json) => {document.getElementById("versionText").innerText = "v" + json["version"];});
+
+        // Display DEBUG
+        Debugger.debugMessage();
     };
 
     return{
@@ -35,23 +37,4 @@ const app = (() => {
 
 addEventListener("DOMContentLoaded", () => {
     app.init();
-    
-    if(DEBUG){
-        fetch('./data/version.json')
-        .then((response) => response.json())
-        .then((json) => {
-            const version = json["version"];
-            const changes = json["changelog"];
-            let msg = "Color Thing by Fırat Usta v" + version + "\nChangelog:";
-            changes.forEach(change => {
-                msg += "\n    -" + change;
-            })
-            console.log(msg);
-    
-            const debugText = document.createElement("p");
-            debugText.classList.add("debugLabel");
-            debugText.innerText = "DEBUG BUILD VERSION v" + version;
-            document.body.appendChild(debugText);
-        });
-    };
 })
